requirements.txt
# Deep Research System — Python Dependencies
# Core
langgraph>=0.2.0
pydantic>=2.0
pyyaml>=6.0

# LLM SDKs
anthropic>=0.39.0
openai>=1.50.0
google-generativeai>=0.8.0

# Storage
sqlalchemy>=2.0
alembic>=1.13
asyncpg>=0.29
redis>=5.0
minio>=7.2

# Observability
prometheus-client>=0.21.0

# Document Generation
python-docx>=1.1.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23.0
pytest-timeout>=2.3.0

# Linting / Type Checking
ruff>=0.7.0
mypy>=1.11
types-PyYAML>=6.0
types-redis>=4.6
